requirements.txt
streamlit
opencv-python-headless==4.10.0.84
mediapipe==0.10.9
numpy
Pillow